+++ packages/node/src/index.ts
@@ -32,10 +32,15 @@ class YogaNodeServer<
   ) {
     super({
       ...options,
+      multipart: options?.multipart !== false,
       fetchAPI:
         options?.fetchAPI ??
         create({
           useNodeFetch: true,
+          formDataLimits:
+            typeof options?.multipart === 'object'
+              ? options.multipart
+              : undefined,
         }),
     })
     this.addressInfo = {