+++ examples/error-handling/CHANGELOG.md
@@ -1,5 +1,26 @@
 # error-masking-example
 
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [8b6d896]
+  - @graphql-yoga/node@2.5.0
+
+## 0.4.1
+
+### Patch Changes
+
+- @graphql-yoga/node@2.4.1
+
+## 0.4.0
+
+### Patch Changes
+
+- Updated dependencies [28e24c3]
+- Updated dependencies [13f96db]
+  - @graphql-yoga/node@2.4.0
+
 ## 0.3.0
 
 ### Patch Changes