+++ examples/graphql-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # graphql-config-example
 
+## 0.5.0
+
+### Patch Changes
+
+- graphql-yoga@2.5.0
+
+## 0.4.1
+
+### Patch Changes
+
+- graphql-yoga@2.4.1
+
+## 0.4.0
+
+### Patch Changes
+
+- graphql-yoga@2.4.0
+
 ## 0.3.0
 
 ### Patch Changes