+++ packages/graphql-yoga/CHANGELOG.md
@@ -1,5 +1,26 @@
 # graphql-yoga
 
+## 2.5.0
+
+### Patch Changes
+
+- Updated dependencies [8b6d896]
+  - @graphql-yoga/node@2.5.0
+
+## 2.4.1
+
+### Patch Changes
+
+- @graphql-yoga/node@2.4.1
+
+## 2.4.0
+
+### Patch Changes
+
+- Updated dependencies [28e24c3]
+- Updated dependencies [13f96db]
+  - @graphql-yoga/node@2.4.0
+
 ## 2.3.0
 
 ### Patch Changes